pyproject.toml
[tool.poetry]
name = "demo_duplicate_counter"
version = "0.1.0"
description = ""
authors = ["Amochkin Alexsandr <aav31415@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10.1"
tornado = "6.1"
pydantic = "1.9.0"
python-dotenv = "0.20.0"
tornado-prometheus = "^0.1.2"
aiopg = "1.3.3"
peewee = "3.14.10"
peewee-async = "^0.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
